config_manager v0.9.5
* _fs_delete Modified: Now checks for iterated file existence BEFORE deletion

~r

+++ config_manager/__init__.py
@@ -78,13 +78,18 @@ def _disable_path_properties(config: dict) -> dict:
 def _fs_delete(directory: Path = None) -> Path:
     if directory.exists():
         for file in directory.iterdir():
-            if file.is_dir():
-                _fs_delete(file)
-                file.rmdir()
-            else:
-                file.unlink()
+
+            if file.exists():
+                if file.is_dir():
+                    _fs_delete(file)
+                    file.rmdir()
+                else:
+                    file.unlink()
+
         directory.rmdir()
+
     else:
+
         print(f'Directory {directory} does not exist')
